build(scripting)!: Rename the module to just "script" for shortness
Also, this will group upcoming modules more nicely.

BREAKING CHANGE: This changes the group of published artifacts,
requiring programmatic users to adapt.
